+++ packages/client/src/tx/types.ts
@@ -38,6 +38,9 @@ export type TxInBestBlocksFound = {
   found: true
 } & TxEventsPayload
 
+export type EventWithTopics = SystemEvent["event"] & {
+  topics: SystemEvent["topics"]
+}
 export type TxEventsPayload = {
   /**
    * Verify if extrinsic was successful, i.e. check if `System.ExtrinsicSuccess`
@@ -48,7 +51,7 @@ export type TxEventsPayload = {
    * Array of all events emitted by the tx. Ordered as they are emitted
    * on-chain.
    */
-  events: Array<SystemEvent["event"]>
+  events: Array<EventWithTopics>
   /**
    * Block information where the tx is found. `hash` of the block, `number` of
    * the block, `index` of the tx in the block.