+++ packages/ink-contracts/src/dynamic-builders.ts
@@ -5,13 +5,19 @@ import {
   Bytes,
   Codec,
   CodecType,
+  createCodec,
   enhanceCodec,
   StringRecord,
   Struct,
   Tuple,
 } from "scale-ts"
 import { InkMetadataLookup } from "./get-lookup"
-import { MessageParamSpec, TypeSpec } from "./metadata-types"
+import {
+  EventParamSpec,
+  EventSpecV5,
+  MessageParamSpec,
+  TypeSpec,
+} from "./metadata-types"
 
 export const getInkDynamicBuilder = (metadataLookup: InkMetadataLookup) => {
   const { metadata } = metadataLookup
@@ -69,20 +75,43 @@ export const getInkDynamicBuilder = (metadataLookup: InkMetadataLookup) => {
     return buildCallable(message)
   }
 
-  const buildEvent = () =>
+  const buildEventCodec = (event: { args: EventParamSpec[] }) =>
+    Struct(
+      Object.fromEntries(
+        event.args.map((param) => [
+          param.label,
+          buildDefinition(param.type.type),
+        ]),
+      ) as StringRecord<Codec<any>>,
+    )
+
+  const buildEvent = (signatureTopic: string | undefined) => {
+    const events = metadata.spec.events as EventSpecV5[]
+
+    const withType = <T>(codec: Codec<T>, type: string) =>
+      enhanceCodec<T, { type: string; value: T }>(
+        codec,
+        (evt) => evt.value,
+        (value) => ({ type, value }),
+      )
+
+    if (signatureTopic) {
+      const event = events.find((evt) => evt.signature_topic === signatureTopic)
+      return event ? withType(buildEventCodec(event), event.label) : null
+    }
+
+    const candidates = events.filter(
+      (evt) => evt.signature_topic === signatureTopic,
+    )
+    return candidates.length
+      ? first(candidates.map((c) => withType(buildEventCodec(c), c.label)))
+      : null
+  }
+
+  const buildEvents = () =>
     Variant(
       Object.fromEntries(
-        metadata.spec.events.map((evt) => [
-          evt.label,
-          Struct(
-            Object.fromEntries(
-              evt.args.map((param) => [
-                param.label,
-                buildDefinition(param.type.type),
-              ]),
-            ) as StringRecord<Codec<any>>,
-          ),
-        ]),
+        metadata.spec.events.map((evt) => [evt.label, buildEventCodec(evt)]),
       ) as StringRecord<Codec<any>>,
     )
 
@@ -90,6 +119,7 @@ export const getInkDynamicBuilder = (metadataLookup: InkMetadataLookup) => {
     buildConstructor,
     buildMessage,
     buildStorage,
+    buildEvents,
     buildEvent,
   }
 }
@@ -108,3 +138,23 @@ const prependBytes = <T>(codec: Codec<T>, hex: string) => {
     ([, value]) => value,
   )
 }
+
+const first = <T>(codecs: Codec<T>[]) =>
+  createCodec<T>(
+    (x) => {
+      for (const codec of codecs) {
+        try {
+          codec.enc(x)
+        } catch (_) {}
+      }
+      throw new Error("Unable to encode")
+    },
+    (x) => {
+      for (const codec of codecs) {
+        try {
+          codec.dec(x)
+        } catch (_) {}
+      }
+      throw new Error("Unable to decode")
+    },
+  )