+++ packages/ink-contracts/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
+## Unreleased
+
 ### Changed
+
+- Dynamic builder and client accept event topics for event decoding.
 
 ## 0.1.0 - 2024-10-12